Able to send transactions now

+++ BLOCKCHAINREST/Blockchain/src/java/blockchain/Blockchain.java
@@ -105,7 +105,7 @@ public class Blockchain extends HttpServlet {
                     JSONObject param = (JSONObject) requestChecker.get("param");
                     String from = (String) param.get("from");
                     String to = (String) param.get("to");
-                    double value = (Double) param.get("value");
+                    long value = (Long) param.get("value");
                     String result = BlockchainHandler.sendTransaction(from, to, value);
                     try (PrintWriter out = response.getWriter()) {
                         out.println(result);
